Reduce number of files
