--- pt_os_web_portal/miniscreen_onboarding/pages/attr/states.py
@@ -1,7 +1,0 @@
-from enum import Enum
-
-
-class RenderState(Enum):
-    STATIONARY = 0
-    ANIMATING = 1
-    DISPLAYING_INFO = 2
